added enter fiscal window and logic

+++ TrackNSave.Server/Models/FormattedReceipt.cs
@@ -3,7 +3,7 @@
     public class FormattedReceipt
     {
         public string User { get; set; } = string.Empty;
-        public int TotalSum { get; set; }
+        public decimal TotalSum { get; set; }
         public string DateTime { get; set; } = string.Empty;
         public string RetailPlace { get; set; } = string.Empty;
         public List<ReceiptItem> Items { get; set; } = new List<ReceiptItem>();
