+++ TrackNSave.Server/Models/ReceiptItem.cs
@@ -3,8 +3,8 @@
     public class ReceiptItem
     {
         public string Name { get; set; } = string.Empty;
-        public int Price { get; set; }
-        public int Quantity { get; set; }
-        public int Sum { get; set; }
+        public decimal Price { get; set; }
+        public decimal Quantity { get; set; }
+        public decimal Sum { get; set; }
     }
 }
